+++ src/utils.cpp
@@ -325,7 +325,7 @@ std::pair<ip, uint16_t> parse_address(const std::string& address_str)
     return std::make_pair(addr, port);
 }
 
-void parse_memory_size(const std::string& value, size_t& out)
+void parse_memory_size(const std::string& value, uint64_t& out)
 {
     enum token
     {
